+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Renaming project to `embeddedstructfieldcheck`.
-- Supporting `suggestedfix` for missing space between embedded types and not embedded types.
+- Supporting `SuggestedFix` for missing space between embedded fields and regular fields.
 
 ## [v0.1.0] 2025-04-24